make:    - исправил проверку стилей. ./test:    - исправил форматирование стилей;    - изменил тесты для s21_sscanf для флага %u 	new file:   .clang-format 	modified:   Makefile 	modified:   s21_sprintf.c 	modified:   s21_sscanf.c 	modified:   s21_string.c 	modified:   s21_string.h 	modified:   s21_utils.c 	modified:   s21_utils.h 	modified:   test.c 	modified:   test/test_s21_sprintf.c 	modified:   test/test_s21_sprintf_c.c 	modified:   test/test_s21_sprintf_d.c 	modified:   test/test_s21_sprintf_e.c 	modified:   test/test_s21_sprintf_f.c 	modified:   test/test_s21_sprintf_g.c 	modified:   test/test_s21_sprintf_o.c 	modified:   test/test_s21_sprintf_p.c 	modified:   test/test_s21_sprintf_pcnt.c 	modified:   test/test_s21_sprintf_s.c 	modified:   test/test_s21_sprintf_u.c 	modified:   test/test_s21_sprintf_x.c 	modified:   test/test_s21_sscanf.c 	modified:   test/test_s21_sscanf_c.c 	modified:   test/test_s21_sscanf_complex.c 	modified:   test/test_s21_sscanf_d.c 	modified:   test/test_s21_sscanf_f.c 	modified:   test/test_s21_sscanf_g.c 	modified:   test/test_s21_sscanf_i.c 	modified:   test/test_s21_sscanf_n.c 	modified:   test/test_s21_sscanf_o.c 	modified:   test/test_s21_sscanf_p.c 	modified:   test/test_s21_sscanf_pcnt.c 	modified:   test/test_s21_sscanf_s.c 	modified:   test/test_s21_sscanf_u.c 	modified:   test/test_s21_sscanf_x.c 	modified:   test/test_s21_string.c 	modified:   test_s21_sscanf_gen.c

+++ src/test/test_s21_sscanf_o.c
@@ -1,8 +1,9 @@
-#include "../s21_string.h"
 #include <check.h>
 #include <limits.h>
 #include <locale.h>
 #include <stdio.h>
+
+#include "../s21_string.h"
 
 START_TEST(test_s21_sscanf_o_simple) {
   const char *input = "01237";
@@ -19,7 +20,7 @@ END_TEST
 
 START_TEST(test_s21_sscanf_ho_simple) {
   const char *input = "012";
-  unsigned x = 10, xr = 10;
+  unsigned short x = 10, xr = 10;
   const char *fmt = "%ho";
 
   int ret_s21 = s21_sscanf(input, fmt, &x);
@@ -59,11 +60,10 @@ END_TEST
 
 START_TEST(test_s21_sscanf_lo_overflow) {
   const char *input = "65554654686452463358438435138143524583451384138432";
-  long unsigned x=0, x_s21=0;
-  // const char *fmt = "%*lo";
+  long unsigned x = 0, x_s21 = 0;
 
-  int ret_s21 = s21_sscanf(input, "%lo",&x_s21);
-  int ret = sscanf(input, "%lo",&x);
+  int ret_s21 = s21_sscanf(input, "%lo", &x_s21);
+  int ret = sscanf(input, "%lo", &x);
 
   ck_assert_int_eq(x, x_s21);
   ck_assert_int_eq(ret, ret_s21);
@@ -253,6 +253,7 @@ END_TEST
 TCase *create_s21_sscanf_o_tests(void) {
   TCase *tc_core_o;
   tc_core_o = tcase_create("Specifier= %o");
+
   tcase_add_test(tc_core_o, test_s21_sscanf_o_simple);
   tcase_add_test(tc_core_o, test_s21_sscanf_ho_simple);
   tcase_add_test(tc_core_o, test_s21_sscanf_lo_simple);
